requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
requests==2.28.2
pydantic==1.10.0
python-dotenv==1.0.1
brother_ql==0.9.4
pdf2image==1.16.0
pyngrok==5.1.0